# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/private/setting": {
            "get": {
                "description": "获取设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "获取设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "key",
                        "name": "key",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "keys",
                        "name": "keys",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "设置",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "保存设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "保存设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "设置",
                        "name": "settings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SettingItem"
                            }
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "删除设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "删除设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/private/setting/token": {
            "post": {
                "description": "重置token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "重置token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/private/settings": {
            "get": {
                "description": "列出设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setting"
                ],
                "summary": "列出设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "group",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "groups",
                        "name": "groups",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "设置列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SettingItem"
                            }
                        }
                    }
                }
            }
        },
        "/api/private/users": {
            "post": {
                "description": "列出用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "列出用户列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "分类信息",
                        "name": "page",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.PageResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.PageResp": {
            "type": "object",
            "properties": {
                "content": {},
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "model.PageReq": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                }
            }
        },
        "model.SettingItem": {
            "type": "object",
            "required": [
                "key"
            ],
            "properties": {
                "flag": {
                    "description": "0 = public, 1 = private, 2 = readonly, 3 = deprecated, etc.",
                    "type": "integer"
                },
                "group": {
                    "description": "use to group setting in frontend",
                    "type": "integer"
                },
                "help": {
                    "description": "help message",
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "key": {
                    "description": "unique key",
                    "type": "string"
                },
                "options": {
                    "description": "values for select",
                    "type": "string"
                },
                "type": {
                    "description": "string, number, bool, select",
                    "type": "string"
                },
                "value": {
                    "description": "value",
                    "type": "string"
                }
            }
        }
    }
}